+++ R/across-learners.R
@@ -1,4 +1,4 @@
-# helper functions
+# helper function that whose arguments correspond to the arguments of GenericML().
 generic.ml.across.learners <- function(Z, D, Y,
                                        propensity_scores,
                                        learners, # need to be mlr3 objects!
@@ -12,6 +12,7 @@ generic.ml.across.learners <- function(Z, D, Y,
                                        vcov_GATES           = setup_vcov(),
                                        equal_variances_CLAN = FALSE,
                                        prop_aux             = 0.5,
+                                       stratify             = setup_stratify(),
                                        quantile_cutoffs     = c(0.25, 0.5, 0.75),
                                        diff_GATES           = setup_diff(),
                                        diff_CLAN            = setup_diff(),
@@ -23,7 +24,14 @@ generic.ml.across.learners <- function(Z, D, Y,
                                        store_learners       = FALSE,
                                        store_splits         = FALSE){
 
-  # call correct main function
+  ## get a function that performs sample splitting by sampling indices for the auxiliary set
+  # either random sampling or stratified sampling, depending on user specification
+  # input checks have been performed in the parent function GenericML()
+  split_fn <- make_split_fn(args_stratified = stratify,
+                            N = nrow(Z),
+                            prop_aux = prop_aux)
+
+  ## call corresponding main function
   do.call(what = get(ifelse(parallel,
                             "generic.ml.across.learners_parallel",
                             "generic.ml.across.learners_serial")),
@@ -31,6 +39,7 @@ generic.ml.across.learners <- function(Z, D, Y,
                       propensity_scores          = propensity_scores,
                       learners                   = learners,
                       learners.names             = learners.names,
+                      split_fn                   = split_fn,
                       num_splits                 = num_splits,
                       Z_CLAN                     = Z_CLAN,
                       X1_BLP                     = X1_BLP,
@@ -58,6 +67,7 @@ generic.ml.across.learners_serial <- function(Z, D, Y,
                                               propensity_scores,
                                               learners, # need to be mlr3 objects!
                                               learners.names,
+                                              split_fn,
                                               num_splits           = 100,
                                               Z_CLAN               = Z,
                                               X1_BLP               = setup_X1(),
@@ -73,10 +83,9 @@ generic.ml.across.learners_serial <- function(Z, D, Y,
                                               significance_level   = 0.05,
                                               min_variation        = 1e-05,
                                               num_cores            = parallel::detectCores(), # dead argument here
-                                              seed                 = NULL,
+                                              seed                 = seed,
                                               store_learners       = FALSE,
                                               store_splits         = FALSE){
-
 
   # set seed
   if(is.null(seed)){
@@ -95,6 +104,7 @@ generic.ml.across.learners_serial <- function(Z, D, Y,
 
   } # IF
 
+  # prepare objects
   num.vars.in.Z_CLAN <- ncol(Z_CLAN)
   genericML.by.split <- list()
   N     <- length(Y)
@@ -117,7 +127,7 @@ generic.ml.across.learners_serial <- function(Z, D, Y,
   for(s in 1:num_splits){
 
     # perform sample splitting into main set and auxiliary set
-    split.ls <- sample_split(D = D, N = N, N_set = N_set, prop = prop)
+    split.ls <- sample_split(split_fn = split_fn, D = D, N = N)
     M_set    <- split.ls$M_set
     A_set    <- split.ls$A_set
 
@@ -183,6 +193,7 @@ generic.ml.across.learners_parallel <- function(Z, D, Y,
                                               propensity_scores,
                                               learners, # need to be mlr3 objects!
                                               learners.names,
+                                              split_fn,
                                               num_splits           = 100,
                                               Z_CLAN               = Z,
                                               X1_BLP               = setup_X1(),
@@ -198,10 +209,9 @@ generic.ml.across.learners_parallel <- function(Z, D, Y,
                                               significance_level   = 0.05,
                                               min_variation        = 1e-05,
                                               num_cores            = parallel::detectCores(),
-                                              seed                 = NULL,
+                                              seed                 = seed,
                                               store_learners       = FALSE,
                                               store_splits         = FALSE){
-
 
   # set seed
   if(is.null(seed)){
@@ -220,6 +230,7 @@ generic.ml.across.learners_parallel <- function(Z, D, Y,
 
   } # IF
 
+  # prepare objects
   num.vars.in.Z_CLAN <- ncol(Z_CLAN)
   K <- length(quantile_cutoffs) + 1
   num.generic_targets.gates <- K + length(diff_GATES$subtracted)
@@ -248,7 +259,7 @@ generic.ml.across.learners_parallel <- function(Z, D, Y,
 
 
     # perform sample splitting into main set and auxiliary set
-    split.ls <- sample_split(D = D, N = N, N_set = N_set, prop = prop)
+    split.ls <- sample_split(split_fn = split_fn, D = D, N = N)
     M_set    <- split.ls$M_set
     A_set    <- split.ls$A_set